Completed creating and showing app schedules in EcheduledAppListView

+++ app/src/main/java/com/interview/appscheduler/core/db/BaseEntity.kt
@@ -1,5 +1,5 @@
 package com.interview.appscheduler.core.db
 
 abstract class BaseEntity {
-    abstract val id: Int
+    abstract val id: Long
 }
